Refactoring to separate Training Strategy from model
